refactor: move contexts out of root
