+++ src/grammar/store.ts
@@ -7,7 +7,7 @@ import {
     AstAsmFunctionDef,
 } from "./ast";
 import { throwInternalCompilerError } from "../error/errors";
-import { CompilerContext, createContextStore } from "../context";
+import { CompilerContext, createContextStore } from "../context/context";
 import { ItemOrigin } from "./src-info";
 import { Parser } from "./grammar";
