+++ src/main.ts
@@ -15,5 +15,5 @@ export { createVirtualFileSystem } from "./vfs/createVirtualFileSystem";
 
 export * from "./browser";
 export * from "./verify";
-export * from "./logger";
+export * from "./context/logger";
 export * from "./error/errors";
