+++ src/types/resolveStatements.spec.ts
@@ -3,7 +3,7 @@ import { resolveDescriptors } from "./resolveDescriptors";
 import { loadCases } from "../utils/loadCases";
 import { openContext } from "../grammar/store";
 import { resolveStatements } from "./resolveStatements";
-import { CompilerContext } from "../context";
+import { CompilerContext } from "../context/context";
 import { featureEnable } from "../config/features";
 import { getParser } from "../grammar";
 import { getAstFactory } from "../grammar/ast";
